refactor: 凝集度を高めるため、新しいクラスを作成
